# Optimize Fraction::fractionize()