This pull request re-implements `fractionize()` on the basis of a [continued fraction expansion][CF] of its argument. These expansions, which have the form


```math
\pi = 3+\frac{1}{7+\frac{1}{15+\frac{1}{1+\frac{1}{292+\ddots}}}}
```

are guaranteed to find all the best rationals approximations of a number, for a specific definition of “best”.

This new implementation is about 50% faster, and more accurate on some arguments. Specifically, it does find the well-known approximation π&nbsp;≈&nbsp;355/113.

Tested on AVR with the following sketch:

<details>

<summary>Test sketch (click to expand/collapse)</summary>

```c++
/*
 * Test the speed and accuracy of float-to-fraction conversions.
 */

#include <fraction.h>

// Numbers to test.
const struct Number {
    const char* name;
    float value;
} numbers[] = {
    { "π ", M_PI },
    { "e ", exp(1) },
    { "φ ", (1 + sqrt(5)) / 2 },
    { "√2", sqrt(2) },
    { nullptr, 0 }  // sentinel
};

void setup() {
    Serial.begin(9600);

    for (int i = 0; numbers[i].name; ++i) {
        const Number& number = numbers[i];

        // Benchmark.
        uint32_t start = micros();
        Fraction f(number.value);
        uint32_t stop = micros();
        float error = number.value - f.toFloat();

        // Show results.
        Serial.print(number.name);
        Serial.print(" = ");
        Serial.print(number.value, 5);
        Serial.print(" = ");
        Serial.print(f.toString());
        if (error != 0) {
            Serial.print(" + ");
            Serial.print(error*1e7);
            Serial.print("e-7");
        }
        Serial.print("  [");
        Serial.print(stop - start);
        Serial.println(" µs]");
    }
}

void loop(){}
```

</details>

The output of the test, in the current master branch:

    π  = 3.14159 = (15689/4994) + 228.88e-7  [1548 µs]
    e  = 2.71828 = (2721/1001)  [1800 µs]
    φ  = 1.61803 = (4181/2584)  [3024 µs]
    √2 = 1.41421 = (3363/2378) + -1.19e-7  [2068 µs]

with this pull-request applied:

    π  = 3.14159 = (355/113) + -2.38e-7  [468 µs]
    e  = 2.71828 = (2721/1001)  [1168 µs]
    φ  = 1.61803 = (10946/6765)  [2072 µs]
    √2 = 1.41421 = (8119/5741)  [1268 µs]

Note that, on π, the error goes down from 96 [ULP][]s to 1 ULP. Also, the golden ratio φ is the worst-case for this algorithm, as it gives the slowest convergence of the expansion.

[CF]: https://en.wikipedia.org/wiki/Simple_continued_fraction
[ULP]: https://en.wikipedia.org/wiki/Unit_in_the_last_place